requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.34.0
starlette>=0.49.1
pydantic==2.11.0
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
prometheus-client==0.21.1
redis==5.2.1
sqlalchemy==2.0.37
psycopg[binary]==3.2.4
pgvector==0.3.6
mcp==1.26.0
pytest==8.3.4
pytest-asyncio==0.25.2
PyJWT==2.10.1
